backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==26.1.0
black==25.1.0
boto3==1.39.12
botocore==1.39.12
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
emergent-plugins @ file:///wheelhouse/emergent_plugins-1.0.0-py3-none-any.whl#sha256=311740f49179f0886b2610d31ad41fc20596a4a9c3980ece847d6305a8a4dd94
emergentintegrations==0.1.1
fastapi==0.104.1
fastuuid==0.14.0
filelock==3.25.2
flake8==7.3.0
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-genai==1.4.0
google-generativeai==0.8.6
googleapis-common-protos==1.70.0
greenlet==3.3.2
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.10
importlib_metadata==9.0.0
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.0.1
jq==1.10.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
litellm @ https://customer-assets.emergentagent.com/internal-asset/library/litellm-1.80.0-py3-none-any.whl#sha256=adf398c513273de9341f61822296c6b2145f7f2dc4a69daf3ac04829f5bde3f8
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
multidict==6.7.1
mypy==1.17.0
mypy_extensions==1.1.0
numpy==2.3.1
oauthlib==3.3.1
openai==1.99.9
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pathspec==0.12.1
pillow==12.1.1
platformdirs==4.3.8
playwright==1.58.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.6
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
pyee==13.0.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.6.0
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.3.32
requests==2.31.0
requests-oauthlib==2.0.0
resend==2.26.0
rich==14.0.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.13.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.27.0
stripe==15.0.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websockets==14.2
yarl==1.23.0
zipp==3.23.0
